refactor: Assign 함수 분리
