+++ app/src/main/res/layout/activity_main.xml
@@ -7,6 +7,8 @@
 
         <import type="android.view.View" />
 
+        <import type="camp.nextstep.edu.tictactoe.AssignBinderKt" />
+
         <variable
             name="viewModel"
             type="camp.nextstep.edu.tictactoe.TicTacToeViewModel" />
